+++ recipe/openid/openidmodels/recipeInterface.go
@@ -15,12 +15,15 @@
 
 package openidmodels
 
-import "github.com/supertokens/supertokens-golang/recipe/jwt/jwtmodels"
+import (
+	"github.com/supertokens/supertokens-golang/recipe/jwt/jwtmodels"
+	"github.com/supertokens/supertokens-golang/supertokens"
+)
 
 type RecipeInterface struct {
-	GetOpenIdDiscoveryConfiguration *func() (GetOpenIdDiscoveryConfigurationResponse, error)
-	CreateJWT                       *func(payload map[string]interface{}, validitySeconds *uint64) (jwtmodels.CreateJWTResponse, error)
-	GetJWKS                         *func() (jwtmodels.GetJWKSResponse, error)
+	GetOpenIdDiscoveryConfiguration *func(userContext supertokens.UserContext) (GetOpenIdDiscoveryConfigurationResponse, error)
+	CreateJWT                       *func(payload map[string]interface{}, validitySeconds *uint64, userContext supertokens.UserContext) (jwtmodels.CreateJWTResponse, error)
+	GetJWKS                         *func(userContext supertokens.UserContext) (jwtmodels.GetJWKSResponse, error)
 }
 
 type GetOpenIdDiscoveryConfigurationResponse struct {