+++ recipe/thirdpartyemailpassword/tpepmodels/apiInterface.go
@@ -17,18 +17,36 @@ package tpepmodels
 
 import (
 	"github.com/supertokens/supertokens-golang/recipe/emailpassword/epmodels"
+	"github.com/supertokens/supertokens-golang/recipe/session/sessmodels"
 	"github.com/supertokens/supertokens-golang/recipe/thirdparty/tpmodels"
+	"github.com/supertokens/supertokens-golang/supertokens"
 )
 
 type APIInterface struct {
-	AuthorisationUrlGET            *func(provider tpmodels.TypeProvider, options tpmodels.APIOptions) (tpmodels.AuthorisationUrlGETResponse, error)
-	AppleRedirectHandlerPOST       *func(code string, state string, options tpmodels.APIOptions) error
-	EmailExistsGET                 *func(email string, options epmodels.APIOptions) (epmodels.EmailExistsGETResponse, error)
-	GeneratePasswordResetTokenPOST *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions) (epmodels.GeneratePasswordResetTokenPOSTResponse, error)
-	PasswordResetPOST              *func(formFields []epmodels.TypeFormField, token string, options epmodels.APIOptions) (epmodels.ResetPasswordUsingTokenResponse, error)
-	ThirdPartySignInUpPOST         *func(provider tpmodels.TypeProvider, code string, authCodeResponse interface{}, redirectURI string, options tpmodels.APIOptions) (ThirdPartyOutput, error)
-	EmailPasswordSignInPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions) (SignInResponse, error)
-	EmailPasswordSignUpPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions) (SignUpResponse, error)
+	AuthorisationUrlGET            *func(provider tpmodels.TypeProvider, options tpmodels.APIOptions, userContext supertokens.UserContext) (tpmodels.AuthorisationUrlGETResponse, error)
+	AppleRedirectHandlerPOST       *func(code string, state string, options tpmodels.APIOptions, userContext supertokens.UserContext) error
+	EmailExistsGET                 *func(email string, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.EmailExistsGETResponse, error)
+	GeneratePasswordResetTokenPOST *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.GeneratePasswordResetTokenPOSTResponse, error)
+	PasswordResetPOST              *func(formFields []epmodels.TypeFormField, token string, options epmodels.APIOptions, userContext supertokens.UserContext) (epmodels.ResetPasswordUsingTokenResponse, error)
+	ThirdPartySignInUpPOST         *func(provider tpmodels.TypeProvider, code string, authCodeResponse interface{}, redirectURI string, options tpmodels.APIOptions, userContext supertokens.UserContext) (ThirdPartyOutput, error)
+	EmailPasswordSignInPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (SignInPOSTResponse, error)
+	EmailPasswordSignUpPOST        *func(formFields []epmodels.TypeFormField, options epmodels.APIOptions, userContext supertokens.UserContext) (SignUpPOSTResponse, error)
+}
+
+type SignUpPOSTResponse struct {
+	OK *struct {
+		User    User
+		Session sessmodels.SessionContainer
+	}
+	EmailAlreadyExistsError *struct{}
+}
+
+type SignInPOSTResponse struct {
+	OK *struct {
+		User    User
+		Session sessmodels.SessionContainer
+	}
+	WrongCredentialsError *struct{}
 }
 
 type EmailpasswordInput struct {
@@ -56,6 +74,7 @@ type ThirdPartyOutput struct {
 		CreatedNewUser   bool
 		User             User
 		AuthCodeResponse interface{}
+		Session          sessmodels.SessionContainer
 	}
 	NoEmailGivenByProviderError *struct{}
 	FieldError                  *struct{ ErrorMsg string }