+++ recipe/thirdpartyemailpassword/tpepmodels/recipeInterface.go
@@ -15,18 +15,21 @@
 
 package tpepmodels
 
-import "github.com/supertokens/supertokens-golang/recipe/emailpassword/epmodels"
+import (
+	"github.com/supertokens/supertokens-golang/recipe/emailpassword/epmodels"
+	"github.com/supertokens/supertokens-golang/supertokens"
+)
 
 type RecipeInterface struct {
-	GetUserByID              *func(userID string) (*User, error)
-	GetUsersByEmail          *func(email string) ([]User, error)
-	GetUserByThirdPartyInfo  *func(thirdPartyID string, thirdPartyUserID string) (*User, error)
-	SignInUp                 *func(thirdPartyID string, thirdPartyUserID string, email EmailStruct) (SignInUpResponse, error)
-	SignUp                   *func(email string, password string) (SignUpResponse, error)
-	SignIn                   *func(email string, password string) (SignInResponse, error)
-	CreateResetPasswordToken *func(userID string) (epmodels.CreateResetPasswordTokenResponse, error)
-	ResetPasswordUsingToken  *func(token string, newPassword string) (epmodels.ResetPasswordUsingTokenResponse, error)
-	UpdateEmailOrPassword    *func(userId string, email *string, password *string) (epmodels.UpdateEmailOrPasswordResponse, error)
+	GetUserByID              *func(userID string, userContext supertokens.UserContext) (*User, error)
+	GetUsersByEmail          *func(email string, userContext supertokens.UserContext) ([]User, error)
+	GetUserByThirdPartyInfo  *func(thirdPartyID string, thirdPartyUserID string, userContext supertokens.UserContext) (*User, error)
+	SignInUp                 *func(thirdPartyID string, thirdPartyUserID string, email EmailStruct, userContext supertokens.UserContext) (SignInUpResponse, error)
+	SignUp                   *func(email string, password string, userContext supertokens.UserContext) (SignUpResponse, error)
+	SignIn                   *func(email string, password string, userContext supertokens.UserContext) (SignInResponse, error)
+	CreateResetPasswordToken *func(userID string, userContext supertokens.UserContext) (epmodels.CreateResetPasswordTokenResponse, error)
+	ResetPasswordUsingToken  *func(token string, newPassword string, userContext supertokens.UserContext) (epmodels.ResetPasswordUsingTokenResponse, error)
+	UpdateEmailOrPassword    *func(userId string, email *string, password *string, userContext supertokens.UserContext) (epmodels.UpdateEmailOrPasswordResponse, error)
 }
 
 type SignInUpResponse struct {